backup/my_pix2pix/requirements.txt
albumentations==1.3.1
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
fonttools==4.42.1
idna==3.4
imageio==2.31.2
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
lazy_loader==0.3
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
nibabel==5.1.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.8.0.76
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.8.25
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
zipp==3.16.2
